Hack fixing S3 pagination issue
When more than a 1000 files are found in an S3 bucket, S3 will paginate results
and multiple requests are necessary to retrieve them all (in which case the
response is said to be truncated). Previously, a check was done on
`isTruncated` and an exception raised if set to true. This commit fixes that
and retrieve all results.

This is considered a hack because the branch on which this is applied is cut
from an old commit of dmlc-core (9416a43). The `master` branch has a bug when
running in distributed mode preventing me from using it.